Split collectors into separate files
