+++ mongo/datadog_checks/mongo/collectors/__init__.py
@@ -1,0 +1,10 @@
+from .base import MongoCollector
+from .coll_stats import CollStatsCollector
+from .current_op import CurrentOpCollector
+from .custom_queries import CustomQueriesCollector
+from .db_stat import DbStatCollector
+from .index_stats import IndexStatsCollector
+from .replica import ReplicaCollector
+from .replication_info import ReplicationInfoCollector
+from .server_status import ServerStatusCollector
+from .top import TopCollector
